% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rsim.R
\name{compute_rsim}
\alias{compute_rsim}
\title{Compute the similarity ratio at each grid point}
\usage{
compute_rsim(m, r_sim, nr, nc)
}
\arguments{
\item{m}{a matrix of size nr x nc storing the occupation information of the house at each grid point.}

\item{r_sim}{a matrix of size nr x nc storing the old similarity ratios at each grid point.}

\item{nr, nc}{the rectangular grid size parameters.}
}
\description{
Compute the similarity ratio at each grid point
}
