% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rsim.R
\name{compute_rsim_1point}
\alias{compute_rsim_1point}
\title{Compute the similarity ratio at the grid point (i,j)}
\usage{
compute_rsim_1point(m, i, j, istart, iend, jstart, jend)
}
\arguments{
\item{m}{a matrix of size nr x nc storing the occupation information of the house at each grid point.}

\item{i, j}{integers: the grid point (i,j) at which the similarity ratio is to be computed.}

\item{istart, iend, jstart, jend}{input parameters that specify the neighbors of (i,j): the neighbors are the grid points at (istart:iend, jstart:jend) excluding (i,j) and locations where m is "white" (empty houses). These parameters are determined by the values of i, j, nr, and nc. They are computed in the functions compute_rsim() and relocate() right before calling this function.}
}
\value{
numeric vector of length 1 returning the similarity ratio at point (i,j).
}
\description{
Compute the similarity ratio at the grid point (i,j)
}
