% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relocate.R
\name{relocate}
\alias{relocate}
\title{Relocate agents to a random empty location}
\usage{
relocate(config, nr, nc, t, r)
}
\arguments{
\item{config}{a list containing two elements named house and rsim. config$house is a nr x nc matrix storing the occupation information of the house at each grid point; config$rsim is a nrxnc matrix storing the similarity ratio at each grid point. This list may be taken directly from the output of the function initial_config() or the output of this function.}

\item{nr, nc}{the rectangular grid size parameters.}

\item{t}{a number between 0 and 1 representing the similarity threshold.}

\item{r}{a number between 0 and 1 specifying the maximum fraction of satisfied agents to relocate.}
}
\value{
A list containing two elements named house and rsim, storing the updated information of config after the relocation of agents.
}
\description{
Relocate agents to a random empty location with similarity ratio greater than or equal to the similarity threshold t.
If no empty locations satisfy the agent's requirement, move the agent to the empty location
with the maximum value of r_sim.
}
