% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_config.R
\name{update_config}
\alias{update_config}
\title{Control the update of configuration and show animation.}
\usage{
update_config(initial, n_agent, nr, nc, t, r, max_iter = 10000L,
  anim_step = 20000L, sleep_time = 0)
}
\arguments{
\item{initial}{the list outputted by the function initial_config().}

\item{n_agent}{an integer representing the total number of agents.}

\item{nr, nc}{the rectangular grid size parameters.}

\item{t}{a number between 0 and 1 representing the similarity threshold.}

\item{r}{a number between 0 and 1 specifying the maximum fraction of satisfied agents to relocate.}

\item{max_iter}{an integer specifying the maximum number of iterations (default is 10000). If the parameter r=0, the simulation will stop when a static equilibrium is reached.}

\item{anim_step}{an integer that that controls the animation: show updated map every anim_step iterations (default is 20000).}

\item{sleep_time}{parameter that controls the animation: pause the execution for sleep_time seconds (default is 0, meaning no pause). This is useful in cases where the execution is too fast for the plots to update.}
}
\description{
Control the update of configuration and show animation.
}
\examples{
nr <- 50; nc <- 50; N <- nr*nc
f_white <- 0.1 # 10\% of unoccupied houses
g <- 4   # 4 groups
fg <- c(1,2,3,4)  # number of agents in the 4 groups has ratios 1:2:3:4
g_color <- c("red","blue","green","gold")
t <- 0.15; r <- 0.05
# set up initial configuration
set.seed(8326454)
initial <- initial_config(nr,nc,f_white, g,fg,g_color)
# Run update_config() for 10 iterations
final <- update_config(initial,n_agent,nr,nc,t,r,10)
}
